Fix conflict with BadgeOS plugin.
We were previously using esc_attr_e() to output the unlisted value.  This
is totally unnecessary as we can just echo the numeric value.

BadgeOS was hooking into the 'gettext' filter greedily, which was triggered
when esc_attr_e() was used.  This caused every creation of achievement
types to be unlisted, which is wrong.

In the future, we should probably use add_post_type_support() to see if
post types should have unlisted support.